- fix update-localtime to not try to copy a timezone file if /etc/localtime   is a symlink
git-svn-id: svn+ssh://thor/srv/subversion/annvix/repos/packages@8041 f191e5f7-4e04-0410-925a-985e7dbff663

+++ releases/2.1-CURRENT/timezone/SOURCES/update-localtime.sh
@@ -11,6 +11,6 @@ if [ -z "${ZONE}" ]; then
     ZONE="Canada/Mountain"
 fi
 
-if [ -f ${zonedir}/${ZONE} ]; then
+if [ -f ${zonedir}/${ZONE} ] && [ ! -L ${localtime_file} ]; then
     cp -f ${zonedir}/${ZONE} ${localtime_file}
 fi
